requirements.txt
openai
python-dotenv
pillow
streamlit
flask